% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCalibration.R
\name{doCalibration}
\alias{doCalibration}
\title{Perform calibration}
\usage{
doCalibration(DF, weights = NULL)
}
\arguments{
\item{DF}{data frame, it must contain a column named 'Concentration' and a column named 'Response'}

\item{weights}{default is NULL}
}
\description{
perform calibration
}
\examples{
\dontrun{
result <- doCalibration(DF)
}
}
\author{
Yonghui Dong
}
