% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doWlm.R
\name{doWlm}
\alias{doWlm}
\title{Perform weighted linear regression}
\usage{
doWlm(DF, weights = NULL)
}
\arguments{
\item{DF}{data frame, it must contain a column named 'Concentration' and a column named 'Response'}

\item{weights}{the weights used in linear regression, default is NULL. User can easily define weights, e.g., "1/x", "1/x^2", "1/y"}
}
\description{
perform weighted linear regression and evaluate by using summed residual
}
\examples{
\dontrun{
result <- doWlm(DF, weights = "1/x^2")
}
}
\author{
Yonghui Dong
}
