% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expData}
\alias{expData}
\title{expData}
\format{
A list with 2 data frames:
\describe{
  \item{noSTD}{the example data without IS}
  \item{STD}{the example data with IS}
}
}
\usage{
expData
}
\description{
Two example data set: one with internal standards (IS), and one without IS
}
\keyword{datasets}
