% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noSTD}
\alias{noSTD}
\title{noSTD}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
  \item{Compound}{Name of the compound}
  \item{Concentration}{Concentration, use 'unknown' to indicate it is a sample type not standard}
  \item{Response}{Response}
}
}
\usage{
noSTD
}
\description{
example data without stable isotope labeled standard
}
\keyword{datasets}
