% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCalibration.R
\name{doCalibration}
\alias{doCalibration}
\title{Perform Calibration}
\usage{
doCalibration(DF, weights = NULL)
}
\arguments{
\item{DF}{data frame, it must contain a column named 'Concentration' and a column named 'Response'}

\item{weights}{default is NULL}
}
\value{
dataframe, the quantification result
}
\description{
Perform calibration
}
\examples{
Concentration <- rep(c(10, 50, 100, "unknown"), each = 3)
Response <- c(133, 156, 177, 6650, 7800, 8850, 13300, 15600, 17700, 156, 1450, 1400)
DF <- cbind.data.frame(Concentration = Concentration, Response = Response)
result <- doCalibration(DF)
}
\author{
Yonghui Dong
}
